LayoutTests/fast/block/block-move-frequent-paint-expected.txt

This tests that we kick into high frequency painting mode properly when animating something quickly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.isPaintingFrequently(square) is false
PASS window.internals.isPaintingFrequently(square) is true
PASS successfullyParsed is true

TEST COMPLETE

